move utils/liquibase to generators/liquibase/support/prepare-field
