+++ generators/bootstrap-application-server/generator.mts
@@ -29,7 +29,7 @@ import {
 } from '../../utils/entity.mjs';
 import type { SpringBootApplication } from '../server/types.mjs';
 import { authenticationTypes, fieldTypes } from '../../jdl/jhipster/index.mjs';
-import { prepareFieldForLiquibaseTemplates } from '../../utils/liquibase.mjs';
+import { prepareField as prepareFieldForLiquibaseTemplates } from '../liquibase/support/index.mjs';
 import { getPomVersionProperties } from '../server/support/index.mjs';
 import { dockerPlaceholderGenerator, getDockerfileContainers } from '../docker/utils.mjs';
 import { GRADLE_VERSION } from '../gradle/constants.mjs';
